requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg[binary]==3.2.5
python-dotenv==1.0.1
sentence-transformers==3.4.1
numpy==2.2.3
pydantic==2.10.6

# PDF extraction
pymupdf==1.24.14